requirements.txt
Flask==2.3.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.5.3
Flask-SocketIO==5.5.1
pymongo==4.6.1
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
eventlet==0.33.3
pandas==2.2.3
reportlab==4.0.4
dnspython==2.3.0
redis==5.0.6
kombu==5.3.4
flask-limiter==2.9
